+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+- Fixed `lf em 410x brute` - better filehandling and memory handling (@iceman1001)
 - Changed split PacketResponseNG status into status and reason(@douniwan5788)
 - add a helper script to decode JEDEC data `script run spi_flash_decode` (@ANTodorov)
 - show SPI flash JEDEC Manufacturer ID and Device ID in `hw status` output (@ANTodorov)